+++ module/op_util.c
@@ -68,7 +68,7 @@ void * rvmalloc(signed long size)
 	adr=(unsigned long) mem;
 	while (size > 0) {
 		page = kvirt_to_pa(adr);
-		mem_map_reserve(virt_to_page(__va(page)));
+		mem_map_reserve(virt_to_page((unsigned long)__va(page)));
 		adr += PAGE_SIZE;
 		size -= PAGE_SIZE;
 	}
@@ -85,7 +85,8 @@ void rvfree(void * mem, signed long size)
 	adr=(unsigned long) mem;
 	while (size > 0) {
 		page = kvirt_to_pa(adr);
-		mem_map_unreserve(virt_to_page(__va(page)));
+		mem_map_unreserve(virt_to_page((unsigned long)__va(page)));
+
 		adr += PAGE_SIZE;
 		size -= PAGE_SIZE;
 	}
